generate session transcript from directPostJWT
