+++ Sources/eudi-lib-ios-wallet-kit/Services/Openid4VpUtils.swift
@@ -17,6 +17,8 @@
 import Foundation
 import SwiftCBOR
 import CryptoKit
+import Logging
+import PresentationExchange
 /**
  *  Utility class to generate the session transcript for the OpenID4VP protocol.
  *
@@ -82,6 +84,31 @@ class Openid4VpUtils {
 		return Data(bytes).base64URLEncodedString()
 	}
 	
+	/// Parse mDoc request from presentation definition (Presentation Exchange 2.0.0 protocol)
+	static func parsePresentationDefinition(_ presentationDefinition: PresentationDefinition, logger: Logger? = nil) throws -> RequestItems? {
+		let pathRx = try NSRegularExpression(pattern: "\\$\\['([^']+)'\\]\\['([^']+)'\\]", options: .caseInsensitive)
+		var res = RequestItems()
+		for inputDescriptor in presentationDefinition.inputDescriptors {
+			guard let fc = inputDescriptor.formatContainer else { logger?.warning("Input descriptor with id \(inputDescriptor.id) is invalid "); continue }
+			guard fc.formats.contains(where: { $0["designation"].string?.lowercased() == "mso_mdoc" }) else { logger?.warning("Input descriptor with id \(inputDescriptor.id) does not contain format mso_mdoc "); continue }
+			let docType = inputDescriptor.id.trimmingCharacters(in: .whitespacesAndNewlines)
+			let kvs: [(String, String)] = inputDescriptor.constraints.fields.compactMap(\.paths.first).compactMap { Self.parsePath($0, pathRx: pathRx) }
+			let nsItems = Dictionary(grouping: kvs, by: \.0).mapValues { $0.map(\.1) }
+			if !nsItems.isEmpty { res[docType] = nsItems }
+		}
+		return res
+	}
+
+	static func parsePath(_ path: String, pathRx: NSRegularExpression) -> (String, String)? {
+		guard let match = pathRx.firstMatch(in: path, options: [], range: NSRange(location: 0, length: path.utf16.count)) else { return nil }
+		let r1 = match.range(at:1);
+		let r1l = path.index(path.startIndex, offsetBy: r1.location)
+		let r1r = path.index(r1l, offsetBy: r1.length)
+		let r2 = match.range(at: 2)
+		let r2l = path.index(path.startIndex, offsetBy: r2.location)
+		let r2r = path.index(r2l, offsetBy: r2.length)
+		return (String(path[r1l..<r1r]), String(path[r2l..<r2r]))
+	}
 	
 }
